+++ RealmSwift/Realm.swift
@@ -110,17 +110,60 @@ public final class Realm {
                                 it will be passed in as an argument.
                                 Otherwise, a `Swift.Error` describing what went wrong will be
                                 passed to the block instead.
+     - returns: A task object which can be used to observe or cancel the async open.
 
      - note: The returned Realm is confined to the thread on which it was created.
              Because GCD does not guarantee that queues will always use the same
              thread, accessing the returned Realm outside the callback block (even if
              accessed from `callbackQueue`) is unsafe.
      */
+    @discardableResult
     public static func asyncOpen(configuration: Realm.Configuration = .defaultConfiguration,
                                  callbackQueue: DispatchQueue = .main,
-                                 callback: @escaping (Realm?, Swift.Error?) -> Void) {
-        RLMRealm.asyncOpen(with: configuration.rlmConfiguration, callbackQueue: callbackQueue) { rlmRealm, error in
+                                 callback: @escaping (Realm?, Swift.Error?) -> Void) -> AsyncOpenTask {
+        return AsyncOpenTask(rlmTask: RLMRealm.asyncOpen(with: configuration.rlmConfiguration, callbackQueue: callbackQueue) { rlmRealm, error in
             callback(rlmRealm.flatMap(Realm.init), error)
+        })
+    }
+
+    /**
+     A task object which can be used to observe or cancel an async open.
+
+     When a synchronized Realm is opened asynchronously, the latest state of the
+     Realm is downloaded from the server before the completion callback is
+     invoked. This task object can be used to observe the state of the download
+     or to cancel it. This should be used instead of trying to observe the
+     download via the sync session as the sync session itself is created
+     asynchronously, and may not exist yet when Realm.asyncOpen() returns.
+     */
+    public struct AsyncOpenTask {
+        fileprivate let rlmTask: RLMAsyncOpenTask
+
+        /**
+         Cancel the asynchronous open.
+
+         Any download in progress will be cancelled, and the completion block for this
+         async open will never be called. If multiple async opens on the same Realm are
+         happening concurrently, all other opens will fail with the error "operation cancelled".
+         */
+        public func cancel() { rlmTask.cancel() }
+
+        /**
+         Register a progress notification block.
+
+         Each registered progress notification block is called whenever the sync
+         subsystem has new progress data to report until the task is either cancelled
+         or the completion callback is called. Progress notifications are delivered on
+         the supplied queue.
+
+         - parameter queue: The queue to deliver progress notifications on.
+         - parameter block: The block to invoke when notifications are available.
+         */
+        public func addProgressNotification(queue: DispatchQueue = .main,
+                                            block: @escaping (SyncSession.Progress) -> Void) {
+            rlmTask.addProgressNotification(on: queue) { transferred, transferrable in
+                block(SyncSession.Progress(transferred: transferred, transferrable: transferrable))
+            }
         }
     }